pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pytest_demo"
version = "0.0.1"
authors = [
  { name="Edward Ji", email="jiziao6@gmail.com" },
]
description = "A small example package to demonstrate pytest"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
"Homepage" = "https://github.com/pypa/sampleproject"
"Bug Tracker" = "https://github.com/pypa/sampleproject/issues"

[project.optional-dependencies]
tests = [
    "allure-pytest",
    "pytest",
]
